pyproject.toml
[tool.poetry]
name = "simple-webhook"
version = "0.1.0"
description = ""
authors = ["RincewindWizzard <git@magierdinge.de>"]
repository = "https://github.com/RincewindWizzard/simple-webhook"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
tomli = "^2.2.1"
python-box = "^7.3.0"
loguru = "^0.7.3"
cryptography = "^44.0.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
simple-webhook = "simple_webhook.__main__:main"